Fixed bug when adding images from plot with symbology
If symbology was applied to the plot, we use mulitple lines for each color.
Since there are multiple lines for multiple subsets of the dataset, we can't
just use event.ind for the index. We instead need to get the X data for the
index of the line picked and match it up to find the correct index